move specific enums in the shared_enums.go file to their specific module declaration. Remove enum declaration that arent directly used in a struct
