+++ pkg/graph/schema.resolvers.go
@@ -17,8 +17,8 @@ import (
 	"github.com/cmsgov/mint-app/pkg/models"
 )
 
-func (r *modelPlanResolver) CmsCenters(ctx context.Context, obj *models.ModelPlan) ([]models.CMSCenter, error) {
-	cmsCenters := models.ConvertEnums[models.CMSCenter](obj.CMSCenters)
+func (r *modelPlanResolver) CmsCenters(ctx context.Context, obj *models.ModelPlan) ([]model.CMSCenter, error) {
+	cmsCenters := models.ConvertEnums[model.CMSCenter](obj.CMSCenters)
 	return cmsCenters, nil
 }
