+++ pkg/models/model_plan.go
@@ -42,3 +42,35 @@ func (m ModelPlan) GetModifiedBy() *string {
 func (m ModelPlan) GetCreatedBy() string {
 	return m.CreatedBy
 }
+
+// ModelCategory represents the category of a model
+type ModelCategory string
+
+// These constants represent the different values of ModelCategory
+const (
+	MCAccountableCare           ModelCategory = "ACCOUNTABLE_CARE"
+	MCDemonstration             ModelCategory = "DEMONSTRATION"
+	MCEBPaymentInitiatives      ModelCategory = "EPISODE_BASED_PAYMENT_INITIATIVES"
+	MCMedicaidAndChip           ModelCategory = "INIT_MEDICAID_CHIP_POP"
+	MCMedicareAndMedicaid       ModelCategory = "INIT__MEDICARE_MEDICAID_ENROLLEES"
+	MCAccelerateDevAndTest      ModelCategory = "INIT_ACCEL_DEV_AND_TEST"
+	MCSpeedBestPracticeAdopt    ModelCategory = "INIT_SPEED_ADOPT_BEST_PRACTICE"
+	MCPrimaryCareTransformation ModelCategory = "PRIMARY_CARE_TRANSFORMATION"
+	MCTBD                       ModelCategory = "UNKNOWN"
+)
+
+// ModelStatus represents the possible statuses of a Model Plan
+type ModelStatus string
+
+// These constants represent the different values of ModelStatus
+const (
+	ModelStatusPlanDraft             ModelStatus = "PLAN_DRAFT"
+	ModelStatusPlanComplete          ModelStatus = "PLAN_COMPLETE"
+	ModelStatusIcipComplete          ModelStatus = "ICIP_COMPLETE"
+	ModelStatusInternalCmmiClearance ModelStatus = "INTERNAL_CMMI_CLEARANCE"
+	ModelStatusCmsClearance          ModelStatus = "CMS_CLEARANCE"
+	ModelStatusHhsClearance          ModelStatus = "HHS_CLEARANCE"
+	ModelStatusOmbAsrfClearance      ModelStatus = "OMB_ASRF_CLEARANCE"
+	ModelStatusCleared               ModelStatus = "CLEARED"
+	ModelStatusAnnounced             ModelStatus = "ANNOUNCED"
+)
