+++ pkg/models/plan_collaborator.go
@@ -18,3 +18,15 @@ type PlanCollaborator struct {
 	ModifiedBy  *string    `json:"modifiedBy" db:"modified_by"`
 	ModifiedDts *time.Time `json:"modifiedDts" db:"modified_dts"`
 }
+
+// TeamRole represents the role of a team member
+type TeamRole string
+
+// These constants represent the different values of TeamRole
+const (
+	TeamRoleModelLead  TeamRole = "MODEL_LEAD"
+	TeamRoleModelTeam  TeamRole = "MODEL_TEAM"
+	TeamRoleLeadership TeamRole = "LEADERSHIP"
+	TeamRoleLearning   TeamRole = "LEARNING"
+	TeamRoleEvaluation TeamRole = "EVALUATION"
+)
