+++ pkg/models/plan_discussion.go
@@ -84,3 +84,13 @@ func (r DiscussionReply) GetModifiedBy() *string {
 func (r DiscussionReply) GetCreatedBy() string {
 	return r.CreatedBy
 }
+
+//DiscussionStatus is an enum that represents the status of a Discussion
+type DiscussionStatus string
+
+//These constants represent the possible values of a DiscussionStatus
+const (
+	DiscussionAnswered   DiscussionStatus = "ANSWERED"
+	DiscussionWaiting    DiscussionStatus = "WAITING_FOR_RESPONSE"
+	DiscussionUnAnswered DiscussionStatus = "UNANSWERED"
+)
